生活福祉資金貸付制度のテストパターンcsvを追加
