+++ make_tests/generate.sh
@@ -1,2 +1,2 @@
 #!/bin/bash
-python3 make_tests.py -i 子育て支援制度テストパターン.csv -c kosodate_teate -o 子育て支援制度テスト
+python3 make_tests.py -i テストパターン/子育て支援制度.csv -c kosodate_teate -o 子育て支援制度テスト
