+++ make_tests/テストパターン/生活福祉資金貸付制度.csv
@@ -1,0 +1,82 @@
+テストID,対象支援制度,親1_年齢,親1_学年,親1_身体障害者手帳等級,親1_療育手帳等級,親1_愛の手帳等級,親1_精神障害者保健福祉手帳等級,親2_年齢,親2_学年,親2_身体障害者手帳等級,親2_療育手帳等級,親2_愛の手帳等級,親2_精神障害者保健福祉手帳等級,第1子_年齢,第1子_学年,第1子_身体障害者手帳等級,第1子_療育手帳等級,第1子_愛の手帳等級,第1子_精神障害者保健福祉手帳等級,住民税非課税世帯,貸付上限額（一意に決まる額 or 最大額）,貸付上限額（最小額 or 一意に決まる場合は空欄）
+1,生活支援費,64,,,,,,,,,,,,,,,,,,FALSE,0,
+2,生活支援費,64,,,,,,,,,,,,,,,,,,TRUE,"200,000",
+3,生活支援費,64,,7,,,,,,,,,,,,,,,,FALSE,"200,000",
+4,生活支援費,64,,,B,,,,,,,,,,,,,,,FALSE,"200,000",
+5,生活支援費,64,,,,4,,,,,,,,,,,,,,FALSE,"200,000",
+6,生活支援費,64,,,,,3,,,,,,,,,,,,,FALSE,"200,000",
+7,生活支援費,65,,,,,,,,,,,,,,,,,,FALSE,"200,000",
+8,生活支援費,64,,7,B,,,,,,,,,,,,,,,FALSE,"200,000",
+9,生活支援費,64,,7,,,,64,,7,,,,,,,,,,FALSE,"200,000",
+10,生活支援費,64,,7,,,,,,,,,,1,,7,,,,FALSE,"200,000",
+11,生活支援費,65,,,,,,65,,,,,,,,,,,,FALSE,"200,000",
+12,生活支援費,64,,,,,,,,,,,,,,,,,,FALSE,"200,000",
+13,生活支援費,65,,,,,,,,,,,,,,,,,,TRUE,"200,000",
+14,一時生活再建費,64,,,,,,,,,,,,,,,,,,FALSE,0,
+15,一時生活再建費,64,,,,,,,,,,,,,,,,,,TRUE,"600,000",
+16,一時生活再建費,64,,7,,,,,,,,,,,,,,,,FALSE,"600,000",
+17,一時生活再建費,64,,,B,,,,,,,,,,,,,,,FALSE,"600,000",
+18,一時生活再建費,64,,,,4,,,,,,,,,,,,,,FALSE,"600,000",
+19,一時生活再建費,64,,,,,3,,,,,,,,,,,,,FALSE,"600,000",
+20,一時生活再建費,65,,,,,,,,,,,,,,,,,,FALSE,"600,000",
+21,一時生活再建費,64,,7,B,,,,,,,,,,,,,,,FALSE,"600,000",
+22,一時生活再建費,64,,7,,,,64,,7,,,,,,,,,,FALSE,"600,000",
+23,一時生活再建費,64,,7,,,,,,,,,,1,,7,,,,FALSE,"600,000",
+24,一時生活再建費,65,,,,,,65,,,,,,,,,,,,FALSE,"600,000",
+25,一時生活再建費,64,,,,,,,,,,,,,,,,,,FALSE,"600,000",
+26,一時生活再建費,65,,,,,,,,,,,,,,,,,,TRUE,"600,000",
+27,福祉費,64,,,,,,,,,,,,,,,,,,FALSE,0,
+28,福祉費,64,,,,,,,,,,,,,,,,,,TRUE,"5,800,000",
+29,福祉費,64,,7,,,,,,,,,,,,,,,,FALSE,"5,800,000",
+30,福祉費,64,,,B,,,,,,,,,,,,,,,FALSE,"5,800,000",
+31,福祉費,64,,,,4,,,,,,,,,,,,,,FALSE,"5,800,000",
+32,福祉費,64,,,,,3,,,,,,,,,,,,,FALSE,"5,800,000",
+33,福祉費,65,,,,,,,,,,,,,,,,,,FALSE,"5,800,000",
+34,福祉費,64,,7,B,,,,,,,,,,,,,,,FALSE,"5,800,000",
+35,福祉費,64,,7,,,,64,,7,,,,,,,,,,FALSE,"5,800,000",
+36,福祉費,64,,7,,,,,,,,,,1,,7,,,,FALSE,"5,800,000",
+37,福祉費,65,,,,,,65,,,,,,,,,,,,FALSE,"5,800,000",
+38,福祉費,64,,,,,,,,,,,,,,,,,,FALSE,"5,800,000",
+39,福祉費,65,,,,,,,,,,,,,,,,,,TRUE,"5,800,000",
+40,緊急小口資金,64,,,,,,,,,,,,,,,,,,FALSE,0,
+41,緊急小口資金,64,,,,,,,,,,,,,,,,,,TRUE,"100,000",
+42,緊急小口資金,64,,7,,,,,,,,,,,,,,,,FALSE,"100,000",
+43,緊急小口資金,64,,,B,,,,,,,,,,,,,,,FALSE,"100,000",
+44,緊急小口資金,64,,,,4,,,,,,,,,,,,,,FALSE,"100,000",
+45,緊急小口資金,64,,,,,3,,,,,,,,,,,,,FALSE,"100,000",
+46,緊急小口資金,65,,,,,,,,,,,,,,,,,,FALSE,"100,000",
+47,緊急小口資金,64,,7,B,,,,,,,,,,,,,,,FALSE,"100,000",
+48,緊急小口資金,64,,7,,,,64,,7,,,,,,,,,,FALSE,"100,000",
+49,緊急小口資金,64,,7,,,,,,,,,,1,,7,,,,FALSE,"100,000",
+50,緊急小口資金,65,,,,,,65,,,,,,,,,,,,FALSE,"100,000",
+51,緊急小口資金,64,,,,,,,,,,,,,,,,,,FALSE,"100,000",
+52,緊急小口資金,65,,,,,,,,,,,,,,,,,,TRUE,"100,000",
+53,住宅入居費,64,,,,,,,,,,,,,,,,,,FALSE,0,
+54,住宅入居費,64,,,,,,,,,,,,,,,,,,TRUE,"400,000",
+55,住宅入居費,64,,7,,,,,,,,,,,,,,,,FALSE,"400,000",
+56,住宅入居費,64,,,B,,,,,,,,,,,,,,,FALSE,"400,000",
+57,住宅入居費,64,,,,4,,,,,,,,,,,,,,FALSE,"400,000",
+58,住宅入居費,64,,,,,3,,,,,,,,,,,,,FALSE,"400,000",
+59,住宅入居費,65,,,,,,,,,,,,,,,,,,FALSE,"400,000",
+60,住宅入居費,64,,7,B,,,,,,,,,,,,,,,FALSE,"400,000",
+61,住宅入居費,64,,7,,,,64,,7,,,,,,,,,,FALSE,"400,000",
+62,住宅入居費,64,,7,,,,,,,,,,1,,7,,,,FALSE,"400,000",
+63,住宅入居費,65,,,,,,65,,,,,,,,,,,,FALSE,"400,000",
+64,住宅入居費,64,,,,,,,,,,,,,,,,,,FALSE,"400,000",
+65,住宅入居費,65,,,,,,,,,,,,,,,,,,TRUE,"400,000",
+66,教育支援費,64,,,,,,,,,,,,,,,,,,FALSE,0,
+67,教育支援費,64,,,,,,,,,,,,,高1,,,,,FALSE,0,
+68,教育支援費,64,,,,,,,,,,,,,高1,,,,,TRUE,"100,000",
+69,教育支援費,64,,,,,,,,,,,,,中3,,,,,TRUE,0,
+70,教育支援費,64,,,,,,,,,,,,,高校生より大きい,,,,,TRUE,"100,000",
+71,教育支援費,64,,,,,,,,,,,,,高校生より大きい,,,,,TRUE,"100,000",
+72,就学支援費,64,,,,,,,,,,,,,,,,,,FALSE,0,
+73,就学支援費,64,,,,,,,,,,,,,中3,,,,,FALSE,0,
+74,就学支援費,64,,,,,,,,,,,,,中3,,,,,TRUE,"500,000",
+75,就学支援費,64,,,,,,,,,,,,,中2,,,,,TRUE,0,
+76,就学支援費,64,,,,,,,,,,,,,高校生より大きい,,,,,TRUE,"500,000",
+77,就学支援費,64,,,,,,,,,,,,,高校生より大きい,,,,,TRUE,"500,000",
+78,不動産担保型生活資金,64,,,,,,,,,,,,,,,,,,FALSE,0,
+79,不動産担保型生活資金,64,,,,,,,,,,,,,,,,,,TRUE,0,
+80,不動産担保型生活資金,65,,,,,,,,,,,,,,,,,,TRUE,"300,000",
+81,不動産担保型生活資金,65,,,,,,65,,,,,,,,,,,,TRUE,"300,000",
